% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{BaseConnection}
\alias{BaseConnection}
\title{connections from the base package
the list of hooks that is related to connections
the names of the hooks
set all the hooks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BaseConnection
}
\arguments{
\item{name}{seet a specific hook given by name, or all the hooks if name=="".
restore all the hooks}

\item{name}{the specific hook with name, or all the hooks if name==""
add a hook
returns a specific saved function}

\item{fun}{the name of the function, preceeded by packagename::
the initializer}
}
\description{
connections from the base package
the list of hooks that is related to connections
the names of the hooks
set all the hooks
}
\keyword{datasets}
