% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{connection.base}
\alias{connection.base}
\title{the hooks related to connections from base}
\format{An object of class \code{BaseConnection} (inherits from \code{R6}) of length 8.}
\usage{
connection.base
}
\description{
the hooks related to connections from base
}
\keyword{datasets}
