% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.handler.R
\docType{data}
\name{FileHandler}
\alias{FileHandler}
\title{this class defines a general interface for handling a file
check if handler can handle the given file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileHandler
}
\arguments{
\item{file}{the file to run}
}
\value{
logical, TRUE if it can run, FALSE if it cannot.
}
\description{
this class defines a general interface for handling a file
check if handler can handle the given file
}
\keyword{datasets}
