% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{fileHook}
\alias{fileHook}
\title{defines a hook for base::file like connection creation functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fileHook
}
\description{
defines a hook for base::file like connection creation functions
}
\keyword{datasets}
