% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{Hook}
\alias{Hook}
\title{Hook class intercept one call in base
subclasses must implement a hook method
the package to hook into
the function that replaces
the saved version of the function
returns the function that this hook intercepts
check if the hook is set
calls the saved function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Hook
}
\arguments{
\item{...}{parameters to be apssed to the original function
set the function hook
restores the functions intercepted by the hooks.
the initializaer}

\item{fun}{the function to intercept}

\item{package}{the package were the function is defined.}
}
\description{
Hook class intercept one call in base
subclasses must implement a hook method
the package to hook into
the function that replaces
the saved version of the function
returns the function that this hook intercepts
check if the hook is set
calls the saved function
}
\keyword{datasets}
