% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.recipe.R
\docType{data}
\name{Interpreter}
\alias{Interpreter}
\title{this class defines an interpreter to run a type of script (distinguished by extensions)
the command to runa script
the command line
the method for making the target from a vector of dependences}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Interpreter
}
\arguments{
\item{script}{the script to run}

\item{target}{the target file}

\item{depend}{the vector of dependences, the first file in depend is the script name
initializer}

\item{pattern}{a list or a vector of patterns that this interpreter can run}

\item{command}{the command to run a given script}

\item{register}{whether toautomatically add to the interpreter manager}
}
\description{
this class defines an interpreter to run a type of script (distinguished by extensions)
the command to runa script
the command line
the method for making the target from a vector of dependences
}
\keyword{datasets}
