% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{is.file}
\alias{is.file}
\title{if the connection is a local file}
\usage{
is.file(con)
}
\arguments{
\item{con}{summary of a connection, as returned by summary.connection}
}
\value{
logical
}
\description{
if the connection is a local file
}
