% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.handler.R
\name{isAbsolutePath}
\alias{isAbsolutePath}
\title{checks if a path is an absolute path}
\usage{
isAbsolutePath(path)
}
\arguments{
\item{path}{the path to check}
}
\value{
logical
}
\description{
checks if a path is an absolute path
}
