% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\docType{data}
\name{Maker}
\alias{Maker}
\title{the Maker class is responsible for making a file.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Maker
}
\arguments{
\item{file}{the file to make}

\item{file}{the file to check
make a file}

\item{file}{the file to make}

\item{force}{force to build the file regardless if it is stale or not.}

\item{silent}{In the case that no rule matches, complain and stop if TRUE, or silently return if FALSE. Still complains and stop if a rule matches but failed to make the file.}

\item{rule}{the rule to add}

\item{replace}{If TRUE, it replaces the rule to make the same target. If FALSE, and a rule to make the same target exists, it complains and fail.
clear the rules
initializer}

\item{dir}{the directory that this maker managers
print the maker}
}
\value{
a MakeRule object, or NULL if do not know how to make it.
check if the file can be handled by this maker

it checks if the file is in the dir (self$pattern)

logical. TRUE if successful, and FALSE if do not know how to make it. If the make fails, the function stops with an error.
add a rule to the list of rules
}
\description{
the Maker class is responsible for making a file.
}
\keyword{datasets}
