% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\docType{data}
\name{MakeTracker}
\alias{MakeTracker}
\title{tracks the files being automatically opened.
a stack of trackers
adding the current file to the top tracker}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MakeTracker
}
\arguments{
\item{file}{the tracked file
starts a new tracker
finish tracking and pop off the stack}
}
\value{
the files being used
}
\description{
tracks the files being automatically opened.
a stack of trackers
adding the current file to the top tracker
}
\keyword{datasets}
