% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.handler.R
\docType{data}
\name{Manager}
\alias{Manager}
\title{defines a file handler manager
the list of file handlers to manage
the class of handlers that it managers
get a handler that can handle a file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Manager
}
\arguments{
\item{file, }{the name of a file to search for a handler}

\item{handler}{a file handler
initializer}
}
\value{
a handler
add  a handler to the first of the list of handlers
}
\description{
defines a file handler manager
the list of file handlers to manage
the class of handlers that it managers
get a handler that can handle a file
}
\keyword{datasets}
