% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath.R
\name{match.stem}
\alias{match.stem}
\title{match file to patterns, which can contain a stem, e.g., dir/\%.c}
\usage{
match.stem(pattern, file)
}
\arguments{
\item{pattern}{the pattern (or a list of vector of them) to match to}

\item{file}{the file name}
}
\value{
logical indicating if file matches pattern, optionally with an attribute named "stem", returning the part of file matching the \%, if the matching pattern contains a \%.
}
\description{
match file to patterns, which can contain a stem, e.g., dir/\%.c
}
