% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.rule.R
\docType{data}
\name{Node}
\alias{Node}
\title{a node in the dependence graph
the dependences
a vector of automatic dependences}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Node
}
\description{
a node in the dependence graph
the dependences
a vector of automatic dependences
}
\keyword{datasets}
