% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath.R
\name{normalizePath}
\alias{normalizePath}
\title{normalize a path}
\usage{
normalizePath(path)
}
\arguments{
\item{path}{the path to normalize}
}
\value{
a vector of normalized path components
}
\description{
normalize a path
}
