% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{openHook}
\alias{openHook}
\title{defines a hook for base::open}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
openHook
}
\description{
defines a hook for base::open
}
\keyword{datasets}
