% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.rule.R
\name{parseTarget}
\alias{parseTarget}
\title{parse a target}
\usage{
parseTarget(target, env)
}
\arguments{
\item{target}{an expression}
}
\value{
a target and its dependences
}
\description{
parse a target
}
