% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{readCharHook}
\alias{readCharHook}
\title{defines a hook for base::readChar
the hook for base::readChar
initializer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
readCharHook
}
\description{
defines a hook for base::readChar
the hook for base::readChar
initializer
}
\keyword{datasets}
