% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.rule.R
\docType{data}
\name{Recipe}
\alias{Recipe}
\title{Recipe is the root class for all recipes for making a file
run is the main function of a recipe for making a target file from a list of dependences}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Recipe
}
\arguments{
\item{target}{the file to make}

\item{depend}{the list of dependences}
}
\value{
this function has no return value. It should throw an error and stop if it fails to make the target.
}
\description{
Recipe is the root class for all recipes for making a file
run is the main function of a recipe for making a target file from a list of dependences
}
\keyword{datasets}
