% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.recipe.R
\docType{data}
\name{RInterpreter}
\alias{RInterpreter}
\title{the interpreter for R scripts
initializer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RInterpreter
}
\arguments{
\item{pattern}{a list or a vector of patterns that this interpreter can run}

\item{register}{whether toautomatically add to the interpreter manager
the method for making the target from a vector of dependences}

\item{script}{the script to run}

\item{target}{the target file}

\item{depend}{the vector of dependences, the first file in depend is the script name}
}
\description{
the interpreter for R scripts
initializer
}
\keyword{datasets}
