% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.handler.R
\docType{data}
\name{Scanner}
\alias{Scanner}
\title{scans for dependences of a file
scan a file for dependences}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Scanner
}
\arguments{
\item{file}{the file to scan}
}
\value{
a vector of dependences, or c() if none
initializer
}
\description{
scans for dependences of a file
scan a file for dependences
}
\keyword{datasets}
