% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.recipe.R
\docType{data}
\name{scriptRecipe}
\alias{scriptRecipe}
\title{this is a Recipe that makes a target file by running a script,
interpreted by an interpreter.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
scriptRecipe
}
\arguments{
\item{target}{the target file}

\item{depend}{the vector of dependences, the first file in depend is the script name
pretty print a scriptRecipe object
initializer}
}
\description{
this is a Recipe that makes a target file by running a script,
interpreted by an interpreter.
}
\keyword{datasets}
