% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{splitPath}
\alias{splitPath}
\title{split a path into a vector of components}
\usage{
splitPath(path)
}
\arguments{
\item{path}{the file path to split}
}
\value{
a vector of path components
}
\description{
split a path into a vector of components
}
