% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.handler.R
\docType{data}
\name{texCompiler}
\alias{texCompiler}
\title{this Interpreter subclass compiles a tex file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
texCompiler
}
\arguments{
\item{ext}{a list or a vector of extensions that this interpreter can run}

\item{register}{whether toautomatically add to the interpreter manager}

\item{latex}{the latex compiler command}

\item{tex}{the plain tex compiler command}

\item{bibtex}{the bibtex processor command
the method for making the target from a vector of dependences}

\item{script}{the script to run}

\item{target}{the target file}

\item{depend}{the vector of dependences, the first file in depend is the script name}
}
\description{
this Interpreter subclass compiles a tex file
}
\keyword{datasets}
