% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.handler.R
\docType{data}
\name{texHandler}
\alias{texHandler}
\title{this class processes a TeX/LaTeX file
the tex file to process
saves the content of the file
matches a command in the form \textbackslash command[]\{\}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
texHandler
}
\arguments{
\item{command}{the tex command to search for}

\item{first.command}{whether this command should be the first command}

\item{to.space}{whether the parameter is up to the first space (in the absence of \{\})
check latex or plain tex}

\item{file}{the tex file to check}
}
\description{
this class processes a TeX/LaTeX file
the tex file to process
saves the content of the file
matches a command in the form \textbackslash command[]\{\}
}
\keyword{datasets}
