% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.handler.R
\docType{data}
\name{texScanner}
\alias{texScanner}
\title{this class defines a TeX/LaTeX file scanner for automatic dependences
scan a file for dependences}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
texScanner
}
\arguments{
\item{file}{the file to scan}
}
\value{
a vector of dependences, or c() if none
initializer
}
\description{
this class defines a TeX/LaTeX file scanner for automatic dependences
scan a file for dependences
}
\keyword{datasets}
