% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_make.R
\name{try_make}
\alias{try_make}
\title{Helper function to make()}
\usage{
try_make(cache.file, builder, verbose = FALSE, fatal = FALSE, ...)
}
\arguments{
\item{cache.file}{The file to be made}

\item{builder}{The builder function}

\item{verbose}{Should I be verbose?}

\item{fatal}{Does something}

\item{...}{Additional parameters to be evaluated in the builder function}
}
\value{
Something
}
\description{
Help do what make() does
}

