% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Make a .rds file}
\usage{
make(cache.file, builder = NULL, verbose = FALSE, max.age = as.difftime(7,
  units = "days"), ...)
}
\arguments{
\item{cache.file}{The file to be made}

\item{builder}{The builder function}

\item{verbose}{Should I be verbose?}

\item{max.age}{If the .rds file exists and is younger than the max.age, skip the builder function}

\item{...}{Additional parameters to be evaluated in the builder function}
}
\value{
An R data object
}
\description{
Makes a .rds file according to a builder function & a recency requirement
}
\examples{

}

