% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{make a file}
\usage{
make(file = "all", force = FALSE)
}
\arguments{
\item{file}{the file to make}

\item{force}{force to build the file regardless if it is stale or not.}
}
\value{
TRUE if successful, FALSE is failed, and NULL if do not know how to make it.
}
\description{
make a file
}
