% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.rule.R
\docType{data}
\name{MakeRule}
\alias{MakeRule}
\title{the base class for make rules
It is the root class for both a node and an implicit rule
It has the notion of a dependence list, a recipeID, and a timestamp
a vector of dependent files
the recipeID for the recipe to make the target
the dependences
the recipeID
the time stamp for last scan, NULL if not scanned
initializer,}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MakeRule
}
\arguments{
\item{depend}{the dependences,}

\item{recipeID}{a unique recipe ID.
pretty print a makeRule object}
}
\description{
the base class for make rules
It is the root class for both a node and an implicit rule
It has the notion of a dependence list, a recipeID, and a timestamp
a vector of dependent files
the recipeID for the recipe to make the target
the dependences
the recipeID
the time stamp for last scan, NULL if not scanned
initializer,
}
\keyword{datasets}
